+++ packages/create-sitecore-jss/src/templates/nextjs-styleguide/scripts/scaffold-component/plugins/next-steps-styleguide.ts
@@ -5,16 +5,10 @@ import { ScaffoldComponentPlugin, ScaffoldComponentPluginConfig } from '..';
  * Set next steps.
  */
 class NextStepsPlugin implements ScaffoldComponentPlugin {
-  order = 100;
+  order = 101;
 
   exec(config: ScaffoldComponentPluginConfig) {
-    const { componentOutputPath, manifestOutputPath, componentName } = config;
-
-    config.nextSteps.push(
-      chalk.green(`
-Scaffolding of ${componentName} complete.
-Next steps:`)
-    );
+    const { manifestOutputPath, componentName } = config;
 
     if (manifestOutputPath) {
       config.nextSteps.push(`* Define the component's data in ${chalk.green(manifestOutputPath)}`);
@@ -24,9 +18,6 @@ Next steps:`)
           `jss deploy component ${componentName} --allowedPlaceholders placeholder-for-component`
         )}, or create the rendering item and datasource template yourself.`
       );
-    }
-    if (componentOutputPath) {
-      config.nextSteps.push(`* Implement the React component in ${chalk.green(componentOutputPath)}`);
     }
     if (manifestOutputPath) {
       config.nextSteps.push(