+++ packages/sitecore-jss-dev-tools/src/templating/index.ts
@@ -1,4 +1,4 @@
-export { ComponentFile, PackageDefinition } from './components';
+export { ComponentFile, PackageDefinition, getComponentList } from './components';
 export { PluginDefinition, generatePlugins, ModuleType } from './plugins';
 export { scaffoldFile } from './scaffold';
 export { strip } from './strip';